This function downloads the metadata for all projects.

Download the metadata from all the projects. This can be useful for finding samples of interests across all projects.

all_metadata(subset = "sra", verbose = TRUE)

Arguments

subset

Either sra, gtex or tcga. Specifies which metadata file to download.

verbose

If TRUE it will print a message of where the file is being downloaded to.

Value

A DataFrame-class object with the phenotype metadata.

Details

Note that for subset = 'gtex', there are more variables than the ones we have for 'sra'. This information corresponds to file GTEx_Data_V6_Annotations_SampleAttributesDS.txt available at http://www.gtexportal.org/home/datasets. There you can find the information describing these variables.

For TCGA we acquired metadata information from 3 different sources:

GDC: via a json query

CGC: via json queries and a custom script to merge the tables

TCGAbiolinks: we used to to parse GDC's XML files For more information, check https://github.com/leekgroup/recount-website/tree/master/metadata/tcga_prep.

Author

Leonardo Collado-Torres

Examples

metadata <- all_metadata()
#> 2023-05-07 01:06:03 downloading the metadata to /tmp/RtmpYvUD3o/metadata_clean_sra.Rdata