Find the GEO accession id for a given SRA run

Given a SRA run id, this function will retrieve the GEO accession id (starting with GSM) if it's available. Otherwise it will return NA.

find_geo(run, verbose = FALSE, sleep = 1/2)

Arguments

run

A character vector of length 1 with the SRA run accession id.

verbose

Whether to print a message for the run. Useful when looping over a larger number of SRA run ids.

sleep

The number of seconds (or fraction) to wait before downloading data using getGEO. This is important if you are looking over geo_info() given the constraints published at https://www.ncbi.nlm.nih.gov/books/NBK25497/.

Value

The GEO accession id for the corresponding sample.

Details

Although the phenotype information already includes the GEO accession ids, not all projects had GEO entries at the time these tables were created. This function will then be useful to check if there is a GEO accession id for a given sample (run). If there is, you can then retrieve the information using geo_info.

Author

Leonardo Collado-Torres

Examples

## Find the GEO accession id for for SRX110461
find_geo("SRX110461")
#> [1] "GSM836270"