Extract information from GEO for a given sample

This function uses GEOquery to extract information for a given sample. The GEO accession ids for the sample can be found in the study phenotype table.

geo_info(
  geoid,
  verbose = FALSE,
  sleep = 1/2,
  getGPL = FALSE,
  destdir = tempdir(),
  ...
)

Arguments

geoid
A character vector of length 1 with the GEO accession id for a given sample.

verbose
If TRUE the geoid will be shown.

sleep
The number of seconds (or fraction) to wait before downloading data using getGEO. This is important if you are looking over geo_info() given the constraints published at https://www.ncbi.nlm.nih.gov/books/NBK25497/.

getGPL
This argument is passed to getGEO and is set to FALSE by default to speed up the process.

destdir
This argument is passed to getGEO.

...
Additional arguments passed to getGEO.

Value

Returns a DataFrame-class with the information from GEO available for the given sample.

Examples

geo_info("GSM836270")

#> Using locally cached version of GSM836270 found here:
#> /tmp/RtmpJ879FJ/GSM836270.soft 

#> DataFrame with 1 row and 25 columns
#>   channel_count data_row_count
#>     <character>    <character>
#> 1             1              0
#>                                                                                                        description
#>                                                                                                        <character>
#> 1 Uninduced K562-shSRF cells used as control to measure the change in mRNA expression profile after SRF knock down
#>   geo_accession
#>     <character>
#> 1     GSM836270
#>                                                                                                               growth_protocol_ch1
#>                                                                                                                       <character>
#> 1 Refer to http://hudsonalpha.org/sites/default/files/pdf/shRNA%20K562%20protocol.pdf for growing conditions for these cell lines
#>               instrument_model last_update_date molecule_ch1 organism_ch1
#>                    <character>      <character>  <character>  <character>
#> 1 Illumina Genome Analyzer IIx      May 15 2019    total RNA Homo sapiens
#>   platform_id source_name_ch1                status submission_date   taxid_ch1
#>   <character>     <character>           <character>     <character> <character>
#> 1    GPL10999          SL2933 Public on Dec 01 2011     Nov 18 2011        9606
#>                                                                                           title
#>                                                                                     <character>
#> 1 K562 cells with shRNA targeting SRF gene cultured with no doxycycline (uninduced - UI), rep1.
#>                                                                                                              treatment_protocol_ch1
#>                                                                                                                         <character>
#> 1 Refer to http://hudsonalpha.org/sites/default/files/pdf/shRNA%20K562%20protocol.pdf for culturing conditions for these cell lines
#>          type                                       characteristics
#>   <character>                                       <CharacterList>
#> 1         SRA cells: K562,shRNA expression: no,treatment: Puromycin
#>     data_processing                           contact
#>     <CharacterList>                   <CharacterList>
#> 1 Bowtie, Cufflinks 601 Genome Way,Huntsville,USA,...
#>                                          extract                     library
#>                                  <CharacterList>             <CharacterList>
#> 1 Standard Illumina RNA-seq library construction cDNA,transcriptomic,RNA-Seq
#>                                                                                                                                                                    relation
#>                                                                                                                                                             <CharacterList>
#> 1 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX110461,BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN00761846,Named Annotation: NA000017789.1 (GSM836270_SL2933.bw)
#>            series
#>   <CharacterList>
#> 1        GSE33816
#>                                                                                                                                                                                                                                                           supplementary_file
#>                                                                                                                                                                                                                                                              <CharacterList>
#> 1 ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM836nnn/GSM836270/suppl/GSM836270_SL2933.bowtie.bam,ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM836nnn/GSM836270/suppl/GSM836270_SL2933.bw,ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM836nnn/GSM836270/suppl/GSM836270_SL2933.gtf.gz