recount: Explore and download data from the recount project

Explore and download data from the recount project available at https://jhubiostatistics.shinyapps.io/recount/. Using the recount package you can download RangedSummarizedExperiment objects at the gene, exon or exon-exon junctions level, the raw counts, the phenotype metadata used, the urls to the sample coverage bigWig files or the mean coverage bigWig file for a particular study. The RangedSummarizedExperiment objects can be used by different packages for performing differential expression analysis. Using http://bioconductor.org/packages/derfinder you can perform annotation-agnostic differential expression analyses with the data from the recount project as described at http://www.nature.com/nbt/journal/v35/n4/full/nbt.3838.html.

See also

Author

Maintainer: Leonardo Collado-Torres lcolladotor@gmail.com (ORCID)

Authors:

Jeffrey T. Leek jtleek@gmail.com (ORCID) [thesis advisor]

Other contributors:

Abhinav Nellore anellore@gmail.com [contributor]

Andrew E. Jaffe andrew.jaffe@libd.org (ORCID) [contributor]

Margaret A. Taub mtaub@jhsph.edu [contributor]

Kai Kammers kai.kammers@gmail.com [contributor]

Shannon E. Ellis sellis18@jhmi.edu (ORCID) [contributor]

Kasper Daniel Hansen kasperdanielhansen@gmail.com (ORCID) [contributor]

Ben Langmead langmea@cs.jhu.edu (ORCID) [contributor]