Search the abstracts from the SRA studies available via the recount project

Given a text query, find the SRA project ids (study accession numbers) that contain the text in their abstract as provided by the SRAdb Bioconductor package.

abstract_search(query, id_only = FALSE, ...)

Arguments

query
A character vector with the text to search for via grep in the abstract info available at recount_abstract.

id_only
Whether to only return the project id or to return summary information for the project(s) that match the query.

...
Additional arguments passed to grep.

Value

If id_only = TRUE it returns a character vector with the project SRA ids (accession numbers). If id_only = FALSE it returns a subset of recount_abstract for the abstracts that contained the query.

Details

Both the query and the abstracts are searched in lower case.

For a more powerful search use the recount project website at https://jhubiostatistics.shinyapps.io/recount/.

See also

Examples

## Find the Geuvadis consortium project
project_info <- abstract_search("Geuvadis consortium")

## See some summary information for this project
project_info

#>    number_samples species
#> 50            667   human
#>                                                                                                                                                                                                                                                                                                abstract
#> 50 The data is under embargo until the first publication by the investigators in early 2013. This RNA sequencing data set of 465 human lymphoblastoid cell line samples from the CEU, FIN, GBR, TSI and YRI populations from the 1000 Genomes sample collection was created by the Geuvadis consortium.
#>      project
#> 50 ERP001942