Query Snaptron to get data from exon-exon junctions present in Intropolis

This function uses the Snaptron API to query specific exon-exon junctions that are available via Intropolis as described in the vignette.

snaptron_query(junctions, version = "srav1", verbose = TRUE, async = TRUE)

Arguments

junctions
A GRanges-class object with the exon-exon junctions of interest. The chromosome names should be in UCSC format, such as 'chr1'. The strand information is ignored in the query.

version
Either srav1, srav2, gtex or tcga. SRA Version 1 of Intropolis has the exon-exon junctions from about 20 thousand RNA-seq samples in hg19 coordinates. SRA Version 2 has the data from about 50 thousand RNA-seq samples aligned to hg38. GTEx has about 30 million junctions from about 10 thousand samples from the GTEx consortium on hg38 coordinates. Finally, TCGA has about 36 million junctions from about 11 thousand samples from the TCGA consortium on hg38 coordinates.

verbose
If TRUE status updates will be printed.

async
Defaults to TRUE but in some situations it might be preferrable to set it to FALSE. This argument gets passed to getURL. Check https://github.com/ChristopherWilks/snaptron/issues/11 for more details.

Value

A GRanges-class object with the results from the Snaptron query. For information on the different columns please see http://snaptron.cs.jhu.edu.

References

Please cite http://snaptron.cs.jhu.edu if you use this function as Snaptron is a separate project from recount. Thank you!

Author

Leonardo Collado-Torres

Examples

library("GenomicRanges")
## Define some exon-exon junctions (hg19 coordinates)
junctions <- GRanges(seqnames = "chr2", IRanges(
    start = c(28971710:28971712, 29555081:29555083, 29754982:29754984),
    end = c(29462417:29462419, 29923338:29923340, 29917714:29917716)
))

## Check against Snaptron SRA version 1 (hg19 coordinates)
snaptron_query(junctions)

#> 2020-12-19 01:24:07 querying Snaptron

#> 2020-12-19 01:26:17 processing results

#> 2020-12-19 01:26:17 extracting information

#> GRanges object with 3 ranges and 14 metadata columns:
#>       seqnames            ranges strand |     type snaptron_id       annotated
#>          <Rle>         <IRanges>  <Rle> | <factor>   <integer> <CharacterList>
#>   [1]     chr2 28971711-29462418      - |  SRAv1:I    22314441            <NA>
#>   [2]     chr2 29555082-29923339      + |  SRAv1:I    22328842            <NA>
#>   [3]     chr2 29754983-29917715      - |  SRAv1:I    22329293               1
#>        left_motif right_motif     left_annotated    right_annotated
#>       <character> <character>    <CharacterList>    <CharacterList>
#>   [1]          AT          AC               <NA>               <NA>
#>   [2]          GC          AG               <NA>               <NA>
#>   [3]          GT          AG aC19,cG19,cG38,... aC19,cG19,cG38,...
#>                  samples read_coverage_by_sample samples_count coverage_sum
#>            <IntegerList>           <IntegerList>     <integer>    <integer>
#>   [1]               3481                       1             1            1
#>   [2] 2643,4867,4945,...               1,1,1,...            13           13
#>   [3]         3,5,13,...               1,1,1,...          1625         5380
#>       coverage_avg coverage_median source_dataset_id
#>          <numeric>       <numeric>         <integer>
#>   [1]        1.000               1                 0
#>   [2]        1.000               1                 0
#>   [3]        3.311               1                 0
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

if (FALSE) {
## Check another set of junctions against SRA version 2 (more data, hg38
## coordinates)
junctions_v2 <- GRanges(seqnames = "chr2", IRanges(
    start = 29532116:29532118, end = 29694848:29694850
))
snaptron_query(junctions_v2, version = "srav2")

## Check these junctions in GTEx and TCGA data
snaptron_query(junctions_v2, version = "gtex")
snaptron_query(junctions_v2, version = "tcga")
}